requirements.txt
streamlit
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langchain-google-genai==1.0.10
langchain-pinecone==0.1.3
sentence-transformers==2.2.2
pypdf
python-dotenv
pinecone>=5.4.2
flask
-e .